Don't update conda first
It should no longer be necessary to force update conda, now that openssl does
its own manual binary prefix replacement in a post-link script. It's better to
not do it if we don't have to so that we can test that old versions of conda
do indeed update themselves correctly (or at least the ones we test in Travis
CI).